requirements.txt
fastapi==0.114.2
uvicorn==0.30.6
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
tldextract==5.1.2
numpy==2.1.1
scikit-learn==1.5.1
pydantic==2.9.2
python-dotenv==1.0.1
sentence-transformers==3.0.1
readability-lxml==0.8.1
lxml_html_clean==0.4.3
